refact: move frontpage-api functions to separate file
