+++ src/types/global.d.ts
@@ -1,5 +1,6 @@
 import DataLoader from 'dataloader';
 import { RequestInit, RequestCache } from 'node-fetch';
+import { FrontpageResponse } from '../data/frontpageApi';
 
 declare global {
   interface AuthToken {
